+++ src/web/assets/ecommercewidget/EcommerceWidgetAsset.php
@@ -10,6 +10,7 @@ namespace dukt\analytics\web\assets\ecommercewidget;
 use craft\web\AssetBundle;
 use craft\web\assets\cp\CpAsset;
 use dukt\analytics\web\assets\analytics\AnalyticsAsset;
+use dukt\analytics\web\assets\googlecharts\GoogleChartsAsset;
 
 /**
  * Realtime report widget asset bundle.
@@ -30,6 +31,7 @@ class EcommerceWidgetAsset extends AssetBundle
         // define the dependencies
         $this->depends = [
             CpAsset::class,
+            GoogleChartsAsset::class,
             AnalyticsAsset::class,
         ];