+++ src/web/assets/realtimereportwidget/RealtimeReportWidgetAsset.php
@@ -11,6 +11,7 @@ use craft\web\AssetBundle;
 use craft\web\assets\cp\CpAsset;
 use craft\web\View;
 use dukt\analytics\web\assets\analytics\AnalyticsAsset;
+use dukt\analytics\web\assets\googlecharts\GoogleChartsAsset;
 
 /**
  * Realtime report widget asset bundle.
@@ -31,6 +32,7 @@ class RealtimeReportWidgetAsset extends AssetBundle
         // define the dependencies
         $this->depends = [
             CpAsset::class,
+            GoogleChartsAsset::class,
             AnalyticsAsset::class,
         ];