+++ src/widgets/Report.php
@@ -137,10 +137,6 @@ class Report extends \craft\base\Widget
                 'cachedResponse' => $cachedResponse ?? null,
             ];
 
-            $view->registerJsFile('//www.gstatic.com/charts/loader.js', [
-                'position' => View::POS_HEAD,
-            ]);
-
             $view->registerAssetBundle(ReportWidgetAsset::class);
 
             $view->registerJs('new Analytics.ReportWidget("widget'.$this->id.'", '.Json::encode($jsOptions).');');